face recognition platform/backendrequirements.txt.txt
flask
flask-cors
python-dotenv
face-recognition
numpy